src/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
uvicorn==0.27.0
asgiref==3.7.2
gunicorn==21.2.0 